test/LayoutTests/Panel/recordMultipleLoads.html

<html>

<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../../../DebugLogger.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_lastChange() {
  console.log("------------test begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/qpd.html"');
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');
  
  var objPropSelector = [
    {type: 'span.cm-string', text:'\"I\'ve been clicked \"'},
    {type: 'span.cm-property', text: 'prop'}
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');

  var getState = 'document.querySelector(".beginRecorded").getAttribute("state")';
  var clickFirstRecorderButton = 'document.querySelector(".beginRecorded").dispatchEvent(new MouseEvent("click"))';
  var clickSecondRecorderButton = 'document.querySelector(".endRecorded").dispatchEvent(new MouseEvent("click"))';
  
  AsyncMachine.pushOp(evaluateInPanel, getState);
  // begin recording
  AsyncMachine.pushOp(evaluateInPanel, clickFirstRecorderButton, function(text) {
    if (text === true)
      return 'record pressed';
  });

  AsyncMachine.pushOp(evaluateInPanel, getState);

  // click twice
  var clickUserButton = 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))';
  AsyncMachine.pushOp(evaluateInPage, clickUserButton);
  AsyncMachine.pushOp(evaluateInPage, clickUserButton);

  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '1:7:1');
  // Stop
  AsyncMachine.pushOp(evaluateInPanel, clickFirstRecorderButton, function(text) {
    if (text === true)
      return 'stop pressed';
  });
  AsyncMachine.pushOp(evaluateInPanel, getState);

  // Replay
  AsyncMachine.pushOp(evaluateInPanel, clickFirstRecorderButton, function(text) {
    if (text === true)
      return 'play pressed';
  });
  AsyncMachine.pushOp(evaluateInPanel, getState);

  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '1:9:3');
  AsyncMachine.pushOp(extractText, 'td.traceValue');

  AsyncMachine.pushOp(clickSelector, '.explainTokenPanel', 'Click to Select Another');

  var conditionValueSelector = [
    {type: 'span.cm-variable', text:'bar'},  // call in onClick
    {type: 'span.cm-variable', text:'bar'},  // func decl
    {type: 'span.cm-property', text: 'value'},  // (!condition.value)
    {type: 'span.cm-property', text: 'value'}   // value  in condition.value
  ];
  AsyncMachine.pushOp(selectTokenInSource, conditionValueSelector);
  AsyncMachine.pushOp(verifyTokenView, 'condition.value');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');  
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '2');
  
  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '2:6');
  AsyncMachine.pushOp(extractText, 'td.traceValue');

  AsyncMachine.pushOp(evaluateInPage, clickUserButton);
  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '2:7');

  AsyncMachine.pushOp(clickSelector, '.explainTokenPanel', 'Click to Select Another');

  var anotherConditionValueSelector = [
    {type: 'span.cm-variable', text:'bar'},
    {type: 'span.cm-variable', text:'bar'},
    {type: 'span.cm-property', text: 'prop'},
    {type: 'span.cm-property', text: 'prop'}  
  ];
  AsyncMachine.pushOp(selectTokenInSource, anotherConditionValueSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');  
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '3');
  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '3:6:0');
  AsyncMachine.pushOp(extractText, 'td.traceValue');

  AsyncMachine.runOp();
}

var testScript = test_lastChange;
</script>
<body>
<h1>Feature Test lastChange</h1>
</body>
</html>
