test/LayoutTests/Panel/recordMultipleLoads-expected.txt

http://localhost:8686/test/qpd.html
prop in   var message = "I've been clicked " + obj.prop + " times";
off
record pressed
record
true
true
stop pressed
recorded
play pressed
recorded
1|0|1|0|1|0|1|0
value in   if (!condition.value) {
undefined|undefined
true
prop in     obj.prop = 0;
1|0|1|0
